AudioRebase: Bring up MS12V2 at Android P [4/7]
PD#SWPL-49855

Problem:
No MS12V2 code at Android P

Solution:
1.Merge audio hal code from Android_R to Android_P and compile success
2.Update dts, dolby decoding library path
3.Delete useless api
4.Add a2dp code -- by Hongchao.Yin

Verify:
t963_ak301

+++ utils/include/aml_ringbuffer.h
@@ -22,7 +22,6 @@ extern "C" {
 #endif
 
 #include <pthread.h>
-#include "aml_malloc_debug.h"
 
 typedef struct ring_buffer {
     pthread_mutex_t lock;
@@ -45,7 +44,7 @@ int ring_buffer_reset_size(struct ring_buffer *rbuffer, int buffer_size);
 int get_buffer_read_space(struct ring_buffer *rbuffer);
 int get_buffer_write_space(struct ring_buffer *rbuffer);
 void ring_buffer_dump(struct ring_buffer *rbuffer);
-size_t ring_buffer_seek(struct ring_buffer *rbuffer, size_t bytes);
+int ring_buffer_seek(struct ring_buffer *rbuffer, int bytes);
 #ifdef __cplusplus
 }
 #endif
